Move StatefulComponent to state package
