+++ src/main/java/org/bric/gui/tabs/WatermarkJPanel.java
@@ -1,6 +1,7 @@
 package org.bric.gui.tabs;
 
 import com.jhlabs.image.RotateFilter;
+import org.bric.gui.state.StatefulComponent;
 import org.bric.gui.swing.JPlacer;
 import org.bric.imageEditParameters.WatermarkParameters;
 import org.bric.utils.Utils;
